experiments folder refactored
