+++ experiments/get_recipes_from_minecraft_data.py
@@ -3,7 +3,7 @@ import minecraft_data
 # pip install minecraft_data
 
 # Java edition minecraft-data
-mcd = minecraft_data("1.11.2")
+mcd = minecraft_data("1.18.2")
 
 
 class MCDataWrapper:
