+++ experiments/test_water.py
@@ -1,8 +1,5 @@
 import cv2
-import common
-import math
 from common import *
-import MalmoPython
 import numpy
 import time
 from examples.neural import get_image
